+++ tourcount/src/main/java/com/wmstein/egm/AbstractMathTransform.kt
@@ -5,7 +5,7 @@ package com.wmstein.egm
  *    GeoTools - The Open Source Java GIS Toolkit
  *    http://geotools.org
  *
- *    (C) 2001-2015, Open Source Geospatial Foundation (OSGeo)
+ *    © 2001-2015, Open Source Geospatial Foundation (OSGeo)
  *
  *    This library is free software; you can redistribute it and/or
  *    modify it under the terms of the GNU Lesser General Public