+++ tourcount/src/main/java/com/wmstein/egm/EarthGravitationalModel.kt
@@ -20,7 +20,7 @@ import kotlin.math.sqrt
  *    GeoTools - The Open Source Java GIS Toolkit
  *    https://geotools.org
  *
- *    (C) 2006-2008, Open Source Geospatial Foundation (OSGeo)
+ *    © 2006-2008, Open Source Geospatial Foundation (OSGeo)
  *
  *    This library is free software; you can redistribute it and/or
  *    modify it under the terms of the GNU Lesser General Public